Refactor code for improved readability and performance
- Simplified return statements in `rand_json` and `rand_obj` functions.
- Consolidated error handling in `invoke_fn`, `invoke_obj`, and `get_obj` functions to reduce nesting.
- Removed unnecessary `if` statements and used `&&` for cleaner conditional checks.
- Enhanced default implementations for `EventContext`, `TriggerExecutionContext`, and `ObjectMetadata` using derive macros.
- Streamlined error handling in `DataService` and `OprcMetaManager` methods.
- Improved clarity in `UnifiedBatchSetHandler` and `UnifiedShardNetwork` by reducing nesting and using more concise patterns.
- Updated health check handler in `oprc-dev-server` to remove commented-out code.
- General code cleanup and formatting adjustments across multiple files for consistency.

+++ commons/oprc-zenoh/src/pool.rs
@@ -65,7 +65,7 @@ impl Pool {
     ) -> Result<zenoh::Session, Box<dyn std::error::Error + Send + Sync>> {
         let mut pool = self.inner.lock().await;
         pool.next_idx += 1;
-        if pool.sessions.len() < self.max_sessions as usize {
+        if pool.sessions.len() < self.max_sessions {
             let mut z_conf = self.z_conf.clone();
             // Assign a unique listen port per session deterministically only when a base port was provided.
             // Otherwise, keep port 0 to let the OS choose a free ephemeral port (collision-free across processes).
